+++ src/components/SearchExamples/SearchExamples.tsx
@@ -57,7 +57,7 @@ interface ISearchExampleProps extends HTMLAttributes<HTMLElement> {
   example: string;
 }
 
-const SearchExample = (props: ISearchExampleProps) => {
+export const SearchExample = (props: ISearchExampleProps) => {
   const { label, example, ...buttonProps } = props;
   return (
     <Grid templateColumns="1fr 2fr" gap={3} my={1}>
@@ -83,28 +83,5 @@ const SearchExample = (props: ISearchExampleProps) => {
         {example}
       </Box>
     </Grid>
-  );
-};
-
-// fallback component to be used as a placeholder
-export const SearchExamplesPlaceholder = () => {
-  return (
-    <Flex justifyContent="center" direction="column" alignItems="center">
-      <Heading as="h3" size="md" mt={3} mb={5}>
-        Search Examples
-      </Heading>
-      <Grid templateColumns={{ base: '1fr', md: 'repeat(2, 1fr)' }} gap={5}>
-        <GridItem>
-          {examples['GENERAL'].left.map(({ label, text }) => (
-            <SearchExample label={label} example={text} key={label} data-text={text} />
-          ))}
-        </GridItem>
-        <GridItem>
-          {examples['GENERAL'].right.map(({ label, text }) => (
-            <SearchExample label={label} example={text} key={label} data-text={text} />
-          ))}
-        </GridItem>
-      </Grid>
-    </Flex>
   );
 };
