+++ src/components/SearchExamples/index.ts
@@ -1,1 +1,2 @@
 export * from './SearchExamples';
+export * from './SearchExamplesPlaceholder';
